+++ app/aurora/[id]/view/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useCharacterView } from './useCharacterView'
+export { useCharacterStats } from './useCharacterStats'